chore: remove duplicated pages components in quests
