+++ apps/govquests-frontend/src/app/quests/[questID]/page.tsx
@@ -1,9 +1,10 @@
 "use client";
 
-import QuestDetails from "@/domains/questing/components/QuestDetails";
-import { useFetchQuest } from "@/domains/questing/hooks/useFetchQuest";
+import React from "react";
 
 import LoadingIndicator from "@/components/ui/LoadingIndicator";
+import { useFetchQuest } from "@/domains/questing/hooks/useFetchQuest";
+import QuestDetails from "@/domains/questing/components/QuestDetails";
 
 interface QuestDetailsPageProps {
   params: {
@@ -16,8 +17,13 @@ export default function QuestDetailsPage({
 }: QuestDetailsPageProps) {
   const { data, isLoading, isError } = useFetchQuest(questID as string);
 
-  if (isLoading) return <LoadingIndicator />;
-  if (isError || !data?.quest) return <p>Error loading quest details.</p>;
+  if (isLoading) {
+    return <LoadingIndicator />;
+  }
+
+  if (isError || !data || !data.quest) {
+    return <p>Error loading quest details.</p>;
+  }
 
   return <QuestDetails quest={data.quest} />;
 }
