+++ apps/govquests-frontend/src/app/quests/page.tsx
@@ -1,14 +1,27 @@
 "use client";
 
-import { useFetchQuests } from "@/domains/questing/hooks/useFetchQuests";
-import QuestList from "@/domains/questing/components/QuestList";
-import LoadingIndicator from "@/components/ui/LoadingIndicator";
+import React from "react";
 
-export default function QuestsPage() {
+import LoadingIndicator from "@/components/ui/LoadingIndicator";
+import QuestList from "@/domains/questing/components/QuestList";
+import { useFetchQuests } from "@/domains/questing/hooks/useFetchQuests";
+
+const QuestsPage: React.FC = () => {
   const { data, isLoading, isError } = useFetchQuests();
 
-  if (isLoading) return <LoadingIndicator />;
-  if (isError || !data) return <p>Error loading quests.</p>;
+  if (isLoading) {
+    return <LoadingIndicator />;
+  }
 
-  return <QuestList quests={data.quests} />;
-}
+  if (isError || !data) {
+    return <p>Error loading quests.</p>;
+  }
+
+  return (
+    <div className="p-8">
+      <QuestList quests={data.quests} />
+    </div>
+  );
+};
+
+export default QuestsPage;
